updated pod proxy according to reviews
