+++ hack/globalscheduler/globalscheduler-up.sh
@@ -107,7 +107,7 @@ do
 done
 
 if [ "x${GO_OUT}" == "x" ]; then
-    make -C "${KUBE_ROOT}" WHAT="cmd/kubectl cmd/hyperkube cmd/kube-apiserver cmd/kubelet cmd/kube-proxy cmd/kube-controller-manager globalscheduler/cmd/gs-controllers globalscheduler/cmd/dispatcher_process  globalscheduler/cmd/distributor_process globalscheduler/cmd/grpc-server cmd/gs-scheduler cmd/resource-collector"
+    make -C "${KUBE_ROOT}" WHAT="cmd/kubectl cmd/hyperkube cmd/kube-apiserver cmd/kubelet cmd/kube-proxy cmd/kube-controller-manager globalscheduler/cmd/gs-controllers globalscheduler/cmd/dispatcher_process  globalscheduler/cmd/distributor_process globalscheduler/cmd/grpc-server globalscheduler/cmd/proxy cmd/gs-scheduler cmd/resource-collector"
 else
     echo "skipped the build."
 fi
@@ -457,6 +457,11 @@ if [[ "${START_MODE}" != "kubeletonly" ]]; then
   kube::common::start_gs_controllers
   echo "Starting grpc server..."
   kube::common::start_grpc_server
+  if [[ "${START_MODE}" != "nokubeproxy" ]]; then
+    kube::common::start_kubeproxy
+  fi
+  echo "Starting pod proxy server..."
+  kube::common::start_proxy_server
   if [[ "${START_MODE}" != "nokubeproxy" ]]; then
     kube::common::start_kubeproxy
   fi
